+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+- Action deployment_update for all 4 supported COE
+
 ## [2.8.0] - 2020-11-16
 
 ### Added